refactor: simplify authorization facade
