+++ app/lib/glossia/extensions.ex
@@ -11,10 +11,6 @@ defmodule Glossia.Extensions do
     Application.get_env(:glossia, :event_handler, Glossia.Events.NoopHandler)
   end
 
-  def authorizer do
-    Application.get_env(:glossia, :authorizer, Glossia.Authorizers.Default)
-  end
-
   def policy_extension do
     Application.get_env(:glossia, :policy_extension, Glossia.PolicyExtensions.Noop)
   end
